LPS-135400 move
