+++ modules/apps/object/object-api/src/main/resources/com/liferay/object/rest/context/path/packageinfo
