+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/instance/lifecycle/SystemObjectDefinitionMetadataPortalInstanceLifecycleListener.java
@@ -15,10 +15,10 @@
 package com.liferay.object.internal.instance.lifecycle;
 
 import com.liferay.object.internal.related.models.SystemObject1toMObjectRelatedModelsProviderImpl;
-import com.liferay.object.internal.rest.context.RESTContextPathResolverImpl;
+import com.liferay.object.internal.rest.context.path.RESTContextPathResolverImpl;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.related.models.ObjectRelatedModelsProvider;
-import com.liferay.object.rest.context.RESTContextPathResolver;
+import com.liferay.object.rest.context.path.RESTContextPathResolver;
 import com.liferay.object.scope.ObjectScopeProviderRegistry;
 import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.object.service.ObjectFieldLocalService;
